[SCRIPTS] Ensure server exist in captureServer...
