+++ packages/scripts/src/js/utilities/captureServerNameFromArg.js
@@ -1,3 +1,5 @@
+const getAvailableServerNames = require('./getAvailableServerNames')
+
 module.exports = function captureServerNameFromArg() {
   const inputArgs = process.argv.slice(2)
   let serverName = ''
@@ -7,6 +9,17 @@ module.exports = function captureServerNameFromArg() {
       serverName = inputArgs[serverNameIndex]
     }
   })
+
+  if (!serverName) {
+    throw new Error(`Server name not provided`)
+  }
+  const availableServerNames = getAvailableServerNames()
+
+  if (!availableServerNames.includes(serverName)) {
+    throw new Error(
+      `Invalid server name: "${serverName}" not found in script env`
+    )
+  }
 
   return serverName
 }
